comp: Add bmp280 sensor
This patch introduces several significant changes to the code.

First and foremost, it adds a submodule to the project containing the
driver for the BMP280 sensor. The driver is authored by Bosch, and no code
changes have been made to it. It has been integrated into the build
system. To compile the project locally, the submodule needs to be
downloaded.

The driver is used as part of a new component. The component has been
tentatively named 'sensor'.

In the main function, sensor initialization and a data read attempt have
been added.

Signed-off-by: Tomasz Leman <apacz87@gmail.com>